Added example. Removed superfluous console app.
