--- SAFE.Console/crust.config
@@ -1,37 +1,0 @@
-{
-"hard_coded_contacts": [
-"178.62.76.8:5483",
-"138.68.185.218:5483",
-"138.68.181.57:5483",
-"138.68.181.60:5483",
-"138.68.181.86:5483",
-"138.68.181.87:5483",
-"138.68.181.168:5483",
-"138.68.181.176:5483",
-"138.68.181.179:5483",
-"138.68.181.180:5483",
-"138.68.181.182:5483",
-"138.68.181.242:5483",
-"138.68.181.243:5483",
-"138.68.181.249:5483",
-"138.68.189.14:5483",
-"138.68.189.15:5483",
-"138.68.189.17:5483",
-"138.68.189.18:5483",
-"138.68.189.19:5483",
-"138.68.189.31:5483",
-"138.68.189.34:5483",
-"138.68.189.36:5483",
-"138.68.189.38:5483",
-"138.68.189.39:5483",
-"46.101.5.179:5483"
-],
-"whitelisted_node_ips": null,
-"whitelisted_client_ips": null,
-"tcp_acceptor_port": 5483,
-"service_discovery_port": null,
-"bootstrap_cache_name": null,
-"force_acceptor_port_in_ext_ep": false,
-"network_name": "alpha_2",
-"dev": null
-}
