--- SAFE.Console/log.toml
@@ -1,39 +1,0 @@
-[appenders.async_console]
-kind = "async_console"
-pattern = "{({l}):1.1} {d(%y-%m-%d %H:%M:%S%.6f)} [{M} #FS#{f}#FE#:{L}] {m}\n"
-
-[[appenders.async_console.filters]]
-kind = "threshold"
-level = "trace"
-
-[appenders.async_file]
-kind = "async_file"
-output_file_name = "Client.log"
-pattern = "{({l}):1.1} {d(%y-%m-%d %H:%M:%S%.6f)} [{M} #FS#{f}#FE#:{L}] {m}\n"
-append = false
-file_timestamp = true
-
-[root]
-level = "error"
-appenders = ["async_console", "async_file"]
-
-[loggers."crust"]
-level = "debug"
-
-[loggers."routing"]
-level = "trace"
-
-[loggers."routing_stats"]
-level = "trace"
-
-[loggers."safe_core"]
-level = "trace"
-
-[loggers."safe_app"]
-level = "trace"
-
-[loggers."safe_authenticator"]
-level = "trace"
-
-[loggers."ffi_utils"]
-level = "trace"
